+++ core/src/lib.rs
@@ -44,8 +44,8 @@ pub mod context;
 pub mod eval;
 pub mod factorize;
 pub mod fmt;
-pub mod gnu_units;
 pub mod load;
+pub mod loader;
 pub mod number;
 pub mod numeric;
 pub mod parsing;
@@ -55,6 +55,7 @@ pub mod substance;
 pub mod types;
 pub mod value;
 
+use loader::gnu_units;
 use parsing::text_query;
 use reply::{QueryError, QueryReply};
